chore: workspace constant language support and refactor
