+++ packages/i18n/src/locales/fr/translations.json
@@ -383,7 +383,7 @@
     "empty_state": {
       "all-issues": {
         "title": "Aucune tâche dans le projet",
-        "description": "Premier projet terminé ! Maintenant, divisez votre travail en morceaux traçables avec des tâches. Allons-y !",
+        "description": "Premier projet terminé ! Maintenant, divisez votre travail en morceaux traçables avec des tâches. Allons-y !",
         "primary_button": {
           "text": "Créer une nouvelle tâche"
         }
@@ -415,6 +415,26 @@
 
   "workspace_settings": {
     "label": "Paramètres de l'espace de travail",
+    "settings": {
+      "general": {
+        "title": "Général"
+      },
+      "members": {
+        "title": "Membres"
+      },
+      "billing-and-plans": {
+        "title": "Facturation et plans"
+      },
+      "exports": {
+        "title": "Exportations"
+      },
+      "webhooks": {
+        "title": "Webhooks"
+      },
+      "api-tokens": {
+        "title": "Jetons API"
+      }
+    },
     "empty_state": {
       "api_tokens": {
         "title": "Aucun jeton API créé",
@@ -440,7 +460,7 @@
     "empty_state": {
       "activity": {
         "title": "Aucune activité pour le moment",
-        "description": "Commencez par créer une nouvelle tâche ! Ajoutez des détails et des propriétés à celle-ci. Explorez davantage Plane pour voir votre activité."
+        "description": "Commencez par créer une nouvelle tâche ! Ajoutez des détails et des propriétés à celle-ci. Explorez davantage Plane pour voir votre activité."
       },
       "assigned": {
         "title": "Aucune tâche assignée",
@@ -507,13 +527,13 @@
   "project_issues": {
     "empty_state": {
       "no_issues": {
-        "title": "Créez un problème et assignez-le à quelqu’un, même à vous-même",
+        "title": "Créez un problème et assignez-le à quelqu'un, même à vous-même",
         "description": "Considérez les problèmes comme des emplois, des tâches, des actions ou JTBD (travaux à accomplir). Nous aimons ça. Un problème et ses sous-problèmes sont généralement des actions basées sur le temps assignées aux membres de votre équipe. Votre équipe crée, assigne et termine des problèmes pour faire avancer votre projet vers son objectif.",
         "primary_button": {
           "text": "Créez votre premier problème",
           "comic": {
             "title": "Les problèmes sont les éléments constitutifs de Plane.",
-            "description": "Redessiner l’interface utilisateur de Plane, Rebrander l’entreprise ou Lancer le nouveau système d’injection de carburant sont des exemples de problèmes qui comportent probablement des sous-problèmes."
+            "description": "Redessiner l'interface utilisateur de Plane, Rebrander l'entreprise ou Lancer le nouveau système d'injection de carburant sont des exemples de problèmes qui comportent probablement des sous-problèmes."
           }
         }
       },
