+++ web/app/[workspaceSlug]/(projects)/settings/sidebar.tsx
@@ -4,13 +4,14 @@ import React from "react";
 import { observer } from "mobx-react";
 import Link from "next/link";
 import { useParams, usePathname } from "next/navigation";
+import { WORKSPACE_SETTINGS_LINKS } from "@plane/constants";
+import { useTranslation } from "@plane/i18n";
 // components
 import { SidebarNavItem } from "@/components/sidebar";
 // hooks
 import { useUserPermissions } from "@/hooks/store";
 // plane web constants
 import { EUserPermissionsLevel } from "@/plane-web/constants/user-permissions";
-import { WORKSPACE_SETTINGS_LINKS } from "@/plane-web/constants/workspace";
 // plane web helpers
 import { shouldRenderSettingLink } from "@/plane-web/helpers/workspace.helper";
 
@@ -19,6 +20,7 @@ export const WorkspaceSettingsSidebar = observer(() => {
   const { workspaceSlug } = useParams();
   const pathname = usePathname();
   // mobx store
+  const { t } = useTranslation();
   const { allowPermissions } = useUserPermissions();
 
   return (
@@ -36,7 +38,7 @@ export const WorkspaceSettingsSidebar = observer(() => {
                     isActive={link.highlight(pathname, `/${workspaceSlug}`)}
                     className="text-sm font-medium px-4 py-2"
                   >
-                    {link.label}
+                    {t(link.i18n_label)}
                   </SidebarNavItem>
                 </Link>
               )
