+++ web/ce/constants/index.ts
@@ -4,4 +4,3 @@ export * from "./gantt-chart";
 export * from "./issues";
 export * from "./project";
 export * from "./user-permissions";
-export * from "./workspace";
