+++ web/core/components/command-palette/actions/workspace-settings-actions.tsx
@@ -4,13 +4,14 @@ import { Command } from "cmdk";
 // hooks
 import Link from "next/link";
 import { useParams } from "next/navigation";
+import { WORKSPACE_SETTINGS_LINKS } from "@plane/constants";
+import { useTranslation } from "@plane/i18n";
 
 // hooks
 import { useUserPermissions } from "@/hooks/store";
 import { useAppRouter } from "@/hooks/use-app-router";
 // plane wev constants
 import { EUserPermissionsLevel } from "@/plane-web/constants/user-permissions";
-import { WORKSPACE_SETTINGS_LINKS } from "@/plane-web/constants/workspace";
 // plane web helpers
 import { shouldRenderSettingLink } from "@/plane-web/helpers/workspace.helper";
 
@@ -25,6 +26,7 @@ export const CommandPaletteWorkspaceSettingsActions: React.FC<Props> = (props) =
   // router params
   const { workspaceSlug } = useParams();
   // mobx store
+  const { t } = useTranslation();
   const { allowPermissions } = useUserPermissions();
   // derived values
 
@@ -47,7 +49,7 @@ export const CommandPaletteWorkspaceSettingsActions: React.FC<Props> = (props) =
               <Link href={`/${workspaceSlug}${setting.href}`}>
                 <div className="flex items-center gap-2 text-custom-text-200">
                   <setting.Icon className="h-4 w-4 text-custom-text-200" />
-                  {setting.label}
+                  {t(setting.i18n_label)}
                 </div>
               </Link>
             </Command.Item>
